added mdx support for blog posts
model: gpt-5.1-codex-max
prompt: create a blog/[slug] route and create a new object data for the
blogs create only 1 blogpost as an example just say Hello world as title
and
  say this is the world in the body. blog object data should consist of
id for the route to redirect, title, date created, mdx file path for the
  body. and yes please try to make if it is possibe to make the  body as
a mdx. no need for author because i will be the only author. id could
  be number or title with dash instead of spacebar i prefer title with
dash.